Split Account class into Account interface and AccountInt implementation.
