+++ src/main/java/com/niccholaspage/Fe/VaultHandler.java
@@ -1,6 +1,7 @@
 package com.niccholaspage.Fe;
 
 import com.niccholaspage.Fe.API.Account;
+import com.niccholaspage.Fe.Databases.AccountInt;
 import java.util.ArrayList;
 import java.util.List;
 import net.milkbowl.vault.economy.Economy;
